feat(studio): global energy heatmap — deviation from unit energy per grid cell
A per-speaker heatmap answers "where does this speaker play"; it cannot
answer "does the renderer conserve energy here". This adds the aggregate
view: total energy over ALL speakers per grid cell, drawn as its
deviation from unit energy — transparent at 0 dB, red above, blue below.
A conserving panner is invisible, so what shows is a real gain or loss:
hot seams between speaker triplets, and the holes left by directions the
layout cannot cover (which is exactly where the out-of-hull modes
differ).

Renderer: BandGaintableFull gains serialize_energy(), emitting the
power-sum amplitude sqrt(sum g^2) per cell per band. Same OBGT container
and the same size as one speaker's slice, so chunking, NACK recovery and
the Tauri decoder are untouched; a GLOBAL_ENERGY_INDEX (-1) subscribe
target selects it. The per-speaker serializer is refactored onto the
shared writer rather than duplicated.

Two signed-ness traps fixed on the way: the Tauri decoder read
speaker_index as u64 (folding the global field onto speaker 0) and the
subscribe command clamped the index with max(0) (erasing the sentinel).

Studio: new scene/global-energy-volume.js maps 20*log10(amplitude) onto a
diverging scale via the volume core's precolored path. It pins the level
normalisation (new optional maxLevel) instead of letting the core divide
by the field's own maximum — without that a 2 dB spread would render
identically to a 20 dB one, which defeats an absolute dB scale.

Toggle, +/-dB scale (1-40, default 6) and band selector in the Display
panel, persisted with the other display prefs, i18n across all 8 locales.
The global and per-speaker heatmaps share one subscription target, so the
global one takes over while enabled.

Verified in the browser against a synthetic table: 0 dB renders fully
transparent, +6 dB red, -6 dB and silence blue, and widening the scale to
12/24 dB drops those same +/-6 dB cells to 0.502/0.251 alpha — linear in
dB and absolute.

+++ omniphony-renderer/orender_engine/src/osc/client_registry.rs
@@ -18,9 +18,11 @@ pub(crate) struct OscClientState {
     /// Version (hash) of the gain table last pushed to this client, so a rebuild
     /// or a re-subscribe carrying the same version skips re-sending the data.
     pub(crate) gaintable_version: Option<u32>,
-    /// Speaker index this client wants the gain table for (the heatmap shows one
-    /// speaker; pushes are serialized per speaker). `None` until first subscribe.
-    pub(crate) gaintable_speaker: Option<usize>,
+    /// What this client wants the gain table for: a speaker index, or
+    /// [`renderer::band_gaintable::GLOBAL_ENERGY_INDEX`] for the all-speaker
+    /// energy field. Pushes are serialized per target. `None` until first
+    /// subscribe.
+    pub(crate) gaintable_speaker: Option<i64>,
 }
 
 pub(crate) struct OscClientRegistry {
@@ -162,15 +164,16 @@ impl OscClientRegistry {
     }
 
     /// Record which speaker a client wants the gain table for.
-    pub(crate) fn set_gaintable_speaker(&self, addr: SocketAddr, speaker: usize) {
+    pub(crate) fn set_gaintable_speaker(&self, addr: SocketAddr, speaker: i64) {
         let mut clients = self.clients.lock().unwrap();
         if let Some(entry) = clients.get_mut(&addr) {
             entry.gaintable_speaker = Some(speaker);
         }
     }
 
-    /// The speaker a client is subscribed for, if any.
-    pub(crate) fn gaintable_speaker(&self, addr: SocketAddr) -> Option<usize> {
+    /// The target a client is subscribed for (speaker index, or
+    /// `GLOBAL_ENERGY_INDEX`), if any.
+    pub(crate) fn gaintable_speaker(&self, addr: SocketAddr) -> Option<i64> {
         self.clients
             .lock()
             .unwrap()
@@ -178,10 +181,10 @@ impl OscClientRegistry {
             .and_then(|c| c.gaintable_speaker)
     }
 
-    /// Live gain-table subscribers as `(addr, last_pushed_version, speaker)`.
+    /// Live gain-table subscribers as `(addr, last_pushed_version, target)`.
     /// Permanent clients always count; timed clients only while within the
     /// heartbeat window.
-    pub(crate) fn gaintable_subscribers(&self) -> Vec<(SocketAddr, Option<u32>, Option<usize>)> {
+    pub(crate) fn gaintable_subscribers(&self) -> Vec<(SocketAddr, Option<u32>, Option<i64>)> {
         let clients = self.clients.lock().unwrap();
         let now = Instant::now();
         clients
